+++ R/MCMC-support.R
@@ -575,7 +575,7 @@ emm_basis.stanreg = function(object, trms, xlev, grid, mode, rescale, ...) {
             }
         }
         
-        misc$respName = as.character(terms(object))[2]
+        misc$respName = as.character.default(terms(object))[2]
     }
     samp = as.matrix(object$stanfit)[, names(bhat), drop = FALSE]
     attr(samp, "n.chains") = object$stanfit@sim$chains